backend/requirements.txt
# BOQ App Backend Dependencies
# AI and Machine Learning
google-generativeai==0.3.2

# Core Flask framework
Flask>=2.3.3,<3.0
Flask-CORS>=4.0.0,<5.0
Werkzeug>=2.3.7,<3.0

# Excel file processing
pandas>=2.2,<3.0
numpy>=2.1,<3.0
openpyxl>=3.1.2,<4.0
xlrd>=2.0.1,<3.0

# HTTP requests for AI API calls
requests>=2.31.0,<3.0

# Data handling and utilities (these are usually pulled by pandas, but harmless to keep)
python-dateutil>=2.8.2,<3.0

# Development and testing
pytest>=7.4.2,<9.0
pytest-flask>=1.2.0,<2.0

# Environment management
python-dotenv>=1.0.0,<2.0

# Logging enhancements
colorlog>=6.7.0,<7.0

# Optional: For production deployment (mostly Linux; fine to leave in)
gunicorn>=21.2.0,<22.0
